+++ src/Pim/Bundle/VersioningBundle/DependencyInjection/PimVersioningExtension.php
@@ -26,7 +26,7 @@ class PimVersioningExtension extends Extension
         $loader->load('managers.yml');
         $loader->load('builders.yml');
         $loader->load('persisters.yml');
-        $loader->load('event_listeners.yml');
+        $loader->load('event_subscribers.yml');
         $loader->load('serializers.yml');
 
         $storageDriver = $container->getParameter('pim_catalog.storage_driver');